moved files from product users to product admin folder
